Move loverushchores and deliveries to types folder
