+++ src/features/island/delivery/components/Orders.tsx
@@ -58,8 +58,8 @@ import { ITEM_IDS } from "features/game/types/bumpkin";
 import { isCollectible } from "features/game/events/landExpansion/garbageSold";
 import { Context } from "features/game/GameProvider";
 import { getActiveCalendarEvent } from "features/game/types/calendar";
-import { getLoveRushStreaks } from "features/game/events/landExpansion/loveRushDeliveries";
-import { getLoveRushDeliveryRewards } from "../../../game/events/landExpansion/loveRushDeliveries";
+import { getLoveRushStreaks } from "features/game/types/loveRushDeliveries";
+import { getLoveRushDeliveryRewards } from "../../../game/types/loveRushDeliveries";
 import { hasFeatureAccess } from "lib/flags";
 import { MachineState } from "features/game/lib/gameMachine";
 import { useSelector } from "@xstate/react";
